+++ next.config.js
@@ -2,7 +2,7 @@ require("newrelic");
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { i18n } = require("./next-i18next.config");
 
-const PROD_BUILD_ID = "production-1.2.5";
+const PROD_BUILD_ID = "production-1.2.6";
 
 module.exports = {
   basePath: "/library-card",