+++ src/components/ApplicationContainer/index.tsx
@@ -1,21 +1,26 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
-import { TemplateAppContainer } from "@nypl/design-system-react-components";
 import { useRouter } from "next/router";
 import React, { useEffect } from "react";
 
 import Banner from "../Banner";
 import ApiErrors from "../ApiErrors";
 import useFormDataContext from "../../context/FormDataContext";
+import {
+  Template,
+  TemplateBreakout,
+  TemplateFull,
+  TemplateHeader,
+  TemplateMain,
+} from "@nypl/design-system-react-components";
 
 const ApplicationContainer = ({ children, problemDetail }) => {
   const {
-    query: { lang = "en" },
+    query: { lang },
   } = useRouter();
   const errorSection = React.createRef<HTMLDivElement>();
   const { state } = useFormDataContext();
   const { errorObj } = state;
   const errorToDisplay = problemDetail ? problemDetail : errorObj;
+  const finalLang = Array.isArray(lang) && lang.length > 0 ? lang[0] : "en";
 
   // If there are errors, focus on the element that displays those errors,
   // for client-side rendering.
@@ -26,22 +31,25 @@ const ApplicationContainer = ({ children, problemDetail }) => {
   }, [errorToDisplay]);
 
   return (
-    <TemplateAppContainer
-      className="nypl-library-card-app"
-      breakout={<Banner />}
-      contentPrimary={
-        <>
-          <ApiErrors
-            lang={lang}
-            ref={errorSection}
-            problemDetail={errorToDisplay}
-          />
-          {children}
-        </>
-      }
-      sidebar="right"
-      marginBottom="l"
-    />
+    <Template>
+      <TemplateHeader>
+        <TemplateBreakout>
+          <Banner />
+        </TemplateBreakout>
+      </TemplateHeader>
+      <TemplateMain id="mainContent">
+        <TemplateFull>
+          <>
+            <ApiErrors
+              lang={finalLang}
+              ref={errorSection}
+              problemDetail={errorToDisplay}
+            />
+            {children}
+          </>
+        </TemplateFull>
+      </TemplateMain>
+    </Template>
   );
 };